+++ source/extensions/lightspeed.trex.properties_pane.shared.mod_setup.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.3]
+### Changed
+- Validation for mod file import
+
 ## [1.5.2]
 ### Changed
 - Update to Kit 106